Small refactor of code, changed name of main class to something simpler, updated tests.
