+++ diorite-utils/injection/src/main/java/org/diorite/inject/utils/Constants.java
@@ -25,7 +25,7 @@
 package org.diorite.inject.utils;
 
 import org.diorite.inject.InjectableClass;
-import org.diorite.inject.InjectionLibrary;
+import org.diorite.inject.Injection;
 import org.diorite.inject.Injector;
 import org.diorite.inject.Qualifiers;
 import org.diorite.inject.controller.ClassData;
@@ -38,7 +38,7 @@ public final class Constants
     public static final TypeDescription.ForLoadedType DEFAULT_INJECTION_CONTROLLER = new TypeDescription.ForLoadedType(DefaultInjectionController.class);
     public static final TypeDescription.ForLoadedType CLASS_DATA                   = new TypeDescription.ForLoadedType(ClassData.class);
     public static final TypeDescription.ForLoadedType INJECTABLE_CLASS             = new TypeDescription.ForLoadedType(InjectableClass.class);
-    public static final TypeDescription.ForLoadedType INJECTION_LIBRARY            = new TypeDescription.ForLoadedType(InjectionLibrary.class);
+    public static final TypeDescription.ForLoadedType INJECTION_LIBRARY            = new TypeDescription.ForLoadedType(Injection.class);
     public static final TypeDescription.ForLoadedType QUALIFIERS                   = new TypeDescription.ForLoadedType(Qualifiers.class);
     public static final TypeDescription.ForLoadedType INSTANCE_SUPPLIER            = new TypeDescription.ForLoadedType(InstanceSupplier.class);
     @SuppressWarnings("deprecation")
